requirements.txt
getch
requests-html
pyyaml
